Add check for Encryption Edge Case where KMS is specified without a specific key.
This edge case for S3 Bucket encryption occurs when KMS encryption is specified for a bucket but no KMS Key ID is selected. When this happens, AWS will default to the AWS KMS Managed Key (aws/s3 KMS Managed Key, not the S3 Managed Key). Thus, there will be no KMSMasterKeyID.

This can occur manually (tested with a CloudFormation deployment with SSEAlgorithm: 'aws:kms' and no KMSMasterKeyID) as well as CDK (Encryption set to KMS_MANAGED).